Build clang after libcxx*
This way our clang will always use our libcxx as soon as possible.
If there are any issues we want things to error out as soon as possible.

As is, it was getting all the way to emerge -e system before clang
was built linking to our libcxx at which point subsequent builds fail.