Add bootstrap_libtapi to build binutils-apple
This "bootstraps" libtapi in stage2 by symlinking the system lib into
EPREFIX/tmp. It also downloads the macports build of libtapi to use
their compiled headers, which allows linking with the system lib.

Once we have a gentoo-built copy of libtapi, we might want to create an
archive with only the headers in it so that we can bootstrap without
downloading the macports build.

Bug: https://bugs.gentoo.org/758167
Signed-off-by: Jacob Floyd <cognifloyd@gmail.com>